+++ nacc_attribute_deriver/attributes/collection/missingness_collection.py
@@ -24,6 +24,7 @@ from nacc_attribute_deriver.utils.constants import (
     MISSINGNESS_VALUES,
 )
 from nacc_attribute_deriver.utils.date import (
+    date_from_form_date,
     find_closest_date,
     standardize_date,
 )
@@ -80,6 +81,9 @@ class FormMissingnessCollection(AttributeCollection):
         )
         self.__prev_record = PreviousRecordNamespace(table=table)
 
+        # if form is associated with a UDS visit (same session)
+        self.__uds_visitdate = date_from_form_date(table.get("_uds_visitdate", None))
+
     @property
     def prev_record(self) -> PreviousRecordNamespace:
         return self.__prev_record
@@ -98,12 +102,15 @@ class FormMissingnessCollection(AttributeCollection):
         Returns:
             visitdate as a string, if found, None otherwise
         """
-        if not date_attribute:
-            date_attribute = self.__form.date_attribute
-            if not date_attribute:
-                return None
+        if date_attribute:
+            raw_visitdate = self.form.get_value(date_attribute, str)
+        elif self.__uds_visitdate:
+            raw_visitdate = self.__uds_visitdate
+        elif self.__form.date_attribute:
+            raw_visitdate = self.form.get_value(self.__form.date_attribute, str)
+        else:
+            return None
 
-        raw_visitdate = self.form.get_value(date_attribute, str)
         if not raw_visitdate:
             raise AttributeDeriverError(
                 f"Unable to find date attribute {date_attribute}"